requirements.txt
alembic==1.14.0
aniso8601==9.0.1
bcrypt==4.2.1
blinker==1.9.0
click==8.1.7
coverage==7.6.9
exceptiongroup==1.2.2
Flask==3.1.0
Flask-Cors==5.0.0
Flask-JWT-Extended==4.7.1
Flask-Migrate==4.0.7
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.8
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.3.4
pytest-cov==6.0.0
pytest-flask==1.3.0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
SQLAlchemy==2.0.36
tomli==2.2.1
typing_extensions==4.12.2
Werkzeug==3.1.3
pytest-mock==3.12.0
